Test that channels are not mixed up by FlowData.write_fcs.
This commit introduces a test which exposes a bug in
FlowData.write_fcs, which means that channels are given the wrong
index in the file created by write_fcs, if there are more than 9
channels.

The issue is caused by the use of the sorted function in write_fcs to
sort the keys of the channel dictionary. This results in a
lexicographical ordering since the dictionary keys are strings (not
integers).